split article and liveblog slot e2e test file
